made edit a new page
